# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Spacenews Flight API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/blogs": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Blogs index",
                "operationId": "9819e5802278e060898795fe3c919599",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/blogs/": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Blogs store",
                "operationId": "717446edb32989e107fd62d672cae614",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "url",
                                    "imageUrl",
                                    "newsSite",
                                    "summary",
                                    "publishedAt",
                                    "updatedAt"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "A blog post :)"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://www.spaceflightnews.com/newpost"
                                    },
                                    "imageUrl": {
                                        "type": "string",
                                        "example": "https://www.spaceflightnews.com/newpost.png"
                                    },
                                    "newsSite": {
                                        "type": "string",
                                        "example": "Space Flight News"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "New post :o"
                                    },
                                    "publishedAt": {
                                        "type": "string",
                                        "example": "2022-05-16T22:43:44.148Z"
                                    },
                                    "updatedAt": {
                                        "type": "string",
                                        "example": "2022-05-16T22:43:44.148Z"
                                    },
                                    "launches": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "l4unc8-1d"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "Launch Provider"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    },
                                    "events": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3v3nt-1d"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "Event Provider"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/blogs/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Blogs show",
                "operationId": "edf2058da1858ee867ebc4789b3195f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Blogs update",
                "operationId": "6a99affa8f7d94cd429d58eb73cf0e88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "url",
                                    "imageUrl",
                                    "newsSite",
                                    "summary",
                                    "publishedAt",
                                    "updatedAt"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Edited blog post :)"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://www.spaceflightnews.com/editedpost"
                                    },
                                    "imageUrl": {
                                        "type": "string",
                                        "example": "https://www.spaceflightnews.com/editedpost.png"
                                    },
                                    "newsSite": {
                                        "type": "string",
                                        "example": "Space Flight News"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "Edited post :o"
                                    },
                                    "publishedAt": {
                                        "type": "string",
                                        "example": "2022-05-16T22:43:44.148Z"
                                    },
                                    "updatedAt": {
                                        "type": "string",
                                        "example": "2022-05-16T22:43:44.148Z"
                                    },
                                    "launches": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "l4unc8-1d"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "Launch Provider"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    },
                                    "events": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "3v3nt-1d"
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "Event Provider"
                                                }
                                            },
                                            "type": "object",
                                            "format": "query"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Blogs destroy",
                "operationId": "9fb0bb3fc65e987a03d2b5ecdda4489d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    }
}